Create SavingsAccountTest.java
program to test class SavingsAccount. Instantiate two savingsAccount objects,
saver1 and saver2, with balances of Rs 2000.00 and Rs 3000.00, respectively. Set
annualInterestRate to 4%, then calculate the monthly interest and print the new balances
for both savers. Then set the annualInterestRate to 5%, calculate the next month's interest
and print the new balances for both savers.